major refactor to utilize same vehicle form
